Split files, add encoder doc comments
